requirements/base.txt
Django==3.1.4
django-crispy-forms==1.10.0
django-environ==0.4.5
psycopg2==2.8.6
celery==5.0.4
